==> 1. Cleaning and Sampling.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

df1=pd.read_csv('/kaggle/input/ids-2018/02-28-2018.csv')
df2=pd.read_csv('/kaggle/input/ids-2018/03-01-2018.csv')
df3=pd.read_csv('/kaggle/input/ids-2018/02-16-2018.csv')
df4=pd.read_csv('/kaggle/input/ids-2018/02-15-2018.csv')
df5=pd.read_csv('/kaggle/input/ids-2018/02-21-2018.csv')
df6=pd.read_csv('/kaggle/input/ids-2018/03-02-2018.csv')
df7=pd.read_csv('/kaggle/input/ids-2018/02-22-2018.csv')
df8=pd.read_csv('/kaggle/input/ids-2018/02-20-2018.csv')
df9=pd.read_csv('/kaggle/input/ids-2018/02-14-2018.csv')
df10=pd.read_csv('/kaggle/input/ids-2018/02-23-2018.csv')

df1.columns = df1.columns.str.strip()
df2.columns = df2.columns.str.strip()
df3.columns = df3.columns.str.strip()
df4.columns = df4.columns.str.strip()
df5.columns = df5.columns.str.strip()
df6.columns = df6.columns.str.strip()
df7.columns = df7.columns.str.strip()
df8.columns = df8.columns.str.strip()
df9.columns = df9.columns.str.strip()
df10.columns = df10.columns.str.strip()

df1=df1.drop(['Dst Port', 'Timestamp'], axis=1)
df2=df2.drop(['Dst Port', 'Timestamp'], axis=1)
df3=df3.drop(['Dst Port', 'Timestamp'], axis=1)
df4=df4.drop(['Dst Port', 'Timestamp'], axis=1)
df5=df5.drop(['Dst Port', 'Timestamp'], axis=1)
df6=df6.drop(['Dst Port', 'Timestamp'], axis=1)
df7=df7.drop(['Dst Port', 'Timestamp'], axis=1)
df8=df8.drop(['Flow ID', 'Src IP', 'Src Port', 'Dst IP', 'Dst Port', 'Timestamp'], axis=1)
df9=df9.drop(['Dst Port', 'Timestamp'], axis=1)
df10=df10.drop(['Dst Port', 'Timestamp'], axis=1)

df1.drop(df1.loc[df1["Label"] == "Label"].index, inplace=True)
df2.drop(df2.loc[df2["Label"] == "Label"].index, inplace=True)
df3.drop(df3.loc[df3["Label"] == "Label"].index, inplace=True)
df4.drop(df4.loc[df4["Label"] == "Label"].index, inplace=True)
df5.drop(df5.loc[df5["Label"] == "Label"].index, inplace=True)
df6.drop(df6.loc[df6["Label"] == "Label"].index, inplace=True)
df7.drop(df7.loc[df7["Label"] == "Label"].index, inplace=True)
df8.drop(df8.loc[df8["Label"] == "Label"].index, inplace=True)
df9.drop(df9.loc[df9["Label"] == "Label"].index, inplace=True)
df10.drop(df10.loc[df10["Label"] == "Label"].index, inplace=True)

for col in df1.columns:
    # Skip the "Label" column
    if col == "Label":
        continue

    # Remove rows where the value in the column is the same as the column name
    df1 = df1[df1[col] != col]

    # Convert the column to int64, coercing errors to NaN if there are any non-numeric values
    df1[col] = pd.to_numeric(df1[col], errors='coerce')

for col in df2.columns:
    # Skip the "Label" column
    if col == "Label":
        continue

    # Remove rows where the value in the column is the same as the column name
    df2 = df2[df2[col] != col]

    # Convert the column to int64, coercing errors to NaN if there are any non-numeric values
    df2[col] = pd.to_numeric(df2[col], errors='coerce')

for col in df3.columns:
    # Skip the "Label" column
    if col == "Label":
        continue

    # Remove rows where the value in the column is the same as the column name
    df3 = df3[df3[col] != col]

    # Convert the column to int64, coercing errors to NaN if there are any non-numeric values
    df3[col] = pd.to_numeric(df3[col], errors='coerce')

df1=df1.replace([np.inf, -np.inf], np.nan)
df2=df2.replace([np.inf, -np.inf], np.nan)
df3=df3.replace([np.inf, -np.inf], np.nan)
df4=df4.replace([np.inf, -np.inf], np.nan)
df5=df5.replace([np.inf, -np.inf], np.nan)
df6=df6.replace([np.inf, -np.inf], np.nan)
df7=df7.replace([np.inf, -np.inf], np.nan)
df8=df8.replace([np.inf, -np.inf], np.nan)
df9=df9.replace([np.inf, -np.inf], np.nan)
df10=df10.replace([np.inf, -np.inf], np.nan)

df1=df1.dropna()
df2=df2.dropna()
df3=df3.dropna()
df4=df4.dropna()
df5=df5.dropna()
df6=df6.dropna()
df7=df7.dropna()
df8=df8.dropna()
df9=df9.dropna()
df10=df10.dropna()

df1=df1.drop_duplicates()
df2=df2.drop_duplicates()
df3=df3.drop_duplicates()
df4=df4.drop_duplicates()
df5=df5.drop_duplicates()
df6=df6.drop_duplicates()
df7=df7.drop_duplicates()
df8=df8.drop_duplicates()
df9=df9.drop_duplicates()
df10=df10.drop_duplicates()

df1=df1.reset_index(drop=True)
df2=df2.reset_index(drop=True)
df3=df3.reset_index(drop=True)
df4=df4.reset_index(drop=True)
df5=df5.reset_index(drop=True)
df6=df6.reset_index(drop=True)
df7=df7.reset_index(drop=True)
df8=df8.reset_index(drop=True)
df9=df9.reset_index(drop=True)
df10=df10.reset_index(drop=True)

print(df1.shape)
print(df2.shape)
print(df3.shape)
print(df4.shape)
print(df5.shape)
print(df6.shape)
print(df7.shape)
print(df8.shape)
print(df9.shape)
print(df10.shape)

df=pd.concat([df1, df2, df3, df4, df5, df6, df7, df8, df9, df10])

df.shape

duplicate_count = df.duplicated().sum()
print(duplicate_count)

df=df.drop_duplicates()

df.shape

label_mapping = {
    'Benign': 'Benign',
    'DDoS attacks-LOIC-HTTP': 'Malicious',
    'DDOS attack-HOIC': 'Malicious',
    'DDOS attack-LOIC-UDP': 'Malicious',
    'DoS attacks-Hulk': 'Malicious',
    'DoS attacks-GoldenEye': 'Malicious',
    'DoS attacks-Slowloris': 'Malicious',
    'DoS attacks-SlowHTTPTest': 'Malicious',
    'Brute Force -Web': 'Malicious',
    'Brute Force -XSS': 'Malicious',
    'SQL Injection': 'Malicious',
    'Bot': 'Malicious',
    'Infilteration': 'Malicious',
    'SSH-Bruteforce': 'Malicious',
    'FTP-BruteForce': 'Malicious'
}

df['Label'] = df['Label'].map(label_mapping)

print(df['Label'].value_counts())

df.to_parquet('CSE-CIC-IDS2018.parquet')

from IPython.display import FileLink
FileLink(r'CSE-CIC-IDS2018.parquet')

sample_df = df.sample(frac=0.2, random_state=42)

sample_df.to_parquet('Sampled Dataset.parquet')

sample_df.shape

print(sample_df['Label'].value_counts())

from IPython.display import FileLink
FileLink(r'Sampled Dataset.parquet')
