==> 2. Scaling and Class Imbalance.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

df=pd.read_parquet('/kaggle/input/cse-cic-ids2018-sampled/Sampled Dataset.parquet')

df['Label'] = df['Label'].map({'Benign': 0, 'Malicious': 1})

from sklearn.preprocessing import MinMaxScaler

num_cols=df.columns[:-1]

min_max_scaler = MinMaxScaler().fit(df[num_cols])

df[num_cols] = min_max_scaler.transform(df[num_cols])

from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split
import joblib

X=df[num_cols]
y=df['Label']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Apply SMOTE
smote = SMOTE(sampling_strategy='auto', random_state=42)
X_train_smote, y_train_smote = smote.fit_resample(X_train, y_train)

joblib.dump((X_train, X_test, y_train, y_test), 'train_test_split.pkl')
joblib.dump((X_train_smote, y_train_smote), 'train_smote.pkl')

from IPython.display import FileLink
FileLink(r'train_smote.pkl')

from IPython.display import FileLink
FileLink(r'train_test_split.pkl')
