==> 3. CNN Feature Extraction.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

import joblib

X_train_smote, y_train_smote = joblib.load('/kaggle/input/scaling-and-smote/train_smote.pkl')
X_test = joblib.load('/kaggle/input/scaling-and-smote/train_test_split.pkl')[1]

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Conv1D, MaxPooling1D, Flatten, Dense

# Reshape the data to 3D for CNN (samples, timesteps, features)
X_train_cnn = X_train_smote.values.reshape((X_train_smote.shape[0], X_train_smote.shape[1], 1))
X_test_cnn = X_test.values.reshape((X_test.shape[0], X_test.shape[1], 1))

# Define CNN model
cnn_model = Sequential()
cnn_model.add(Conv1D(filters=64, kernel_size=3, activation='relu', input_shape=(X_train_cnn.shape[1], 1)))
cnn_model.add(MaxPooling1D(pool_size=2))
cnn_model.add(Flatten())
cnn_model.add(Dense(64, activation='relu'))
cnn_model.add(Dense(1, activation='sigmoid'))

cnn_model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
cnn_model.fit(X_train_cnn, y_train_smote, epochs=10, batch_size=32)

# Extract features using CNN
cnn_features = cnn_model.predict(X_train_cnn)
cnn_features_test = cnn_model.predict(X_test_cnn)

np.save('cnn_features_train.npy', cnn_features)
np.save('cnn_features_cnn_test.npy', cnn_features_test)
features = {'train': cnn_features, 'test': cnn_features_test}
np.save('cnn_features.npy', features, allow_pickle=True)
cnn_model.save('cnn_model.h5')

from IPython.display import FileLink
FileLink(r'cnn_features_train.npy')

from IPython.display import FileLink
FileLink(r'cnn_features_cnn_test.npy')

from IPython.display import FileLink
FileLink(r'cnn_features.npy')

from IPython.display import FileLink
FileLink(r'cnn_model.h5')
