==> 3. LSTM Feature Extraction.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

import joblib

X_train_smote, y_train_smote = joblib.load('/kaggle/input/scaling-and-smote/train_smote.pkl')
X_test = joblib.load('/kaggle/input/scaling-and-smote/train_test_split.pkl')[1]

##### import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import LSTM, Dense
from tensorflow.keras.optimizers import Adam

# Reshape the data to 3D for LSTM (samples, timesteps, features)
X_train_lstm = X_train_smote.values.reshape((X_train_smote.shape[0], X_train_smote.shape[1], 1))
X_test_lstm = X_test.values.reshape((X_test.shape[0], X_test.shape[1], 1))

# Define LSTM model
lstm_model = Sequential()
lstm_model.add(LSTM(64, input_shape=(X_train_lstm.shape[1], 1)))
lstm_model.add(Dense(64, activation='relu'))
lstm_model.add(Dense(1, activation='sigmoid'))

optimizer = Adam(learning_rate=0.001, clipnorm=1.0)
lstm_model.compile(optimizer=optimizer, loss='binary_crossentropy', metrics=['accuracy'])
lstm_model.fit(X_train_lstm, y_train_smote, epochs=10, batch_size=32)

# Extract features using LSTM
lstm_features = lstm_model.predict(X_train_lstm)
lstm_features_test = lstm_model.predict(X_test_lstm)

np.save('lstm_features_train.npy', lstm_features)
np.save('lstm_features_test.npy', lstm_features_test)
features = {'train': lstm_features, 'test': lstm_features_test}
np.save('lstm_features.npy', features, allow_pickle=True)
lstm_model.save('lstm_model.h5')

from IPython.display import FileLink
FileLink(r'lstm_features_train.npy')

from IPython.display import FileLink
FileLink(r'lstm_features_test.npy')

from IPython.display import FileLink
FileLink(r'lstm_features.npy')

from IPython.display import FileLink
FileLink(r'lstm_model.h5')
