==> 3. SAE Feature Extraction.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

import joblib

X_train_smote = joblib.load('/kaggle/input/scaling-and-smote/train_smote.pkl')[0]
X_test = joblib.load('/kaggle/input/scaling-and-smote/train_test_split.pkl')[1]

from tensorflow.keras.layers import Input, Dense
from tensorflow.keras.models import Model

# Define Autoencoder architecture
input_dim = X_train_smote.shape[1]
input_layer = Input(shape=(input_dim,))
encoded = Dense(128, activation='relu')(input_layer)
encoded = Dense(64, activation='relu')(encoded)
encoded = Dense(32, activation='relu')(encoded)

decoded = Dense(64, activation='relu')(encoded)
decoded = Dense(128, activation='relu')(decoded)
decoded = Dense(input_dim, activation='sigmoid')(decoded)

autoencoder = Model(input_layer, decoded)
autoencoder.compile(optimizer='adam', loss='mse')

# Train Autoencoder
autoencoder.fit(X_train_smote, X_train_smote, epochs=10, batch_size=32)

# Extract features using encoder part of Autoencoder
encoder = Model(input_layer, encoded)
autoencoder_features = encoder.predict(X_train_smote)
autoencoder_features_test = encoder.predict(X_test)

np.save('sae_features_train.npy', autoencoder_features)
np.save('sae_features_test.npy', autoencoder_features_test)
features = {'train': autoencoder_features, 'test': autoencoder_features_test}
np.save('sae_features.npy', features, allow_pickle=True)
autoencoder.save('autoencoder.h5')
encoder.save('encoder.h5')

from IPython.display import FileLink
FileLink(r'sae_features_train.npy')

from IPython.display import FileLink
FileLink(r'sae_features_test.npy')

from IPython.display import FileLink
FileLink(r'cnn_features.npy')

from IPython.display import FileLink
FileLink(r'autoencoder.h5')

from IPython.display import FileLink
FileLink(r'encoder.h5')
